VOAE-22 Use a pool of write tokens
The Brightcove API requires us to provide a write token for each upload
performed, and will throw an error if two simultaneous uploads try to share a
write token.

This change modifies the BrightcoveMediaService to work with a pool of write
tokens, ensuring that each concurrent update uses a different write token.