NOJIRA Have the MediaNode record whether each version is ready to play or not.
This saves having to query the media service for a video's upload status each
time it gets played.  This is important because the Brightcove query for upload
status is classified as a write operation, and requires a write token.  Since
write tokens are a scarce resource, we want to minimise contention for them.