VOAE-12 Refactor the MediaService to provide a File instead of an InputStream.
Since we now control the upload process with our FileUploadFilter, we know
we'll have access to a real underlying file.  This makes our MediaService
implementations much simpler, since we can easily get the file size in bytes
and as many FileInputStreams as we like, without havinag to worry about
reading the same stream multiple times.